+++ lib/bones_api_db_gcp.dart
@@ -1,0 +1,4 @@
+/// Bones_API DB Adapters for Google Cloud Platform (GCP).
+library bones_api_adapter.gcp;
+
+export 'src/bones_api_entity_db_object_gcs.dart';